qemu: jdwp: fix UB
Shifting left by more than the number of bits in the LHS operand is UB.

Fixes the failing test found via compiler upgrade.

Link: https://wiki.sei.cmu.edu/confluence/display/c/INT34-C.+Do+not+shift+an+expression+by+a+negative+number+of+bits+or+by+greater+than+or+equal+to+the+number+of+bits+that+exist+in+the+operand
Bug: 147705683
Test: run external\qemu\android\android-emu\android\jdwp\Jdwp_unittest.cpp

+++ android/android-emu/android/jdwp/Jdwp_unittest.cpp
@@ -21,12 +21,12 @@ using namespace android::jdwp;
 TEST(Jdwp, ReadBuffer) {
     const uint8_t kByteBuffer[8] = {0, 0, 0, 0, 0, 0, 0, 0x1};
     EXPECT_EQ(0x1, readValFromBuffer<uint32_t>(kByteBuffer + 4, 4));
-    EXPECT_EQ(0x1, readValFromBuffer<uint32_t>(kByteBuffer, 8)); // fail case
+    EXPECT_EQ(0x1, readValFromBuffer<uint64_t>(kByteBuffer, 8));
     EXPECT_EQ(0x1, uint32FromBuffer(kByteBuffer + 4));
 
     const uint8_t kByteBuffer2[8] = {0, 0, 0, 0, 0, 0, 0, 0x2};
     EXPECT_EQ(0x2, readValFromBuffer<uint32_t>(kByteBuffer2 + 4, 4));
-    EXPECT_EQ(0x2, readValFromBuffer<uint32_t>(kByteBuffer2, 8)); // fail case
+    EXPECT_EQ(0x2, readValFromBuffer<uint64_t>(kByteBuffer2, 8));
     EXPECT_EQ(0x2, uint32FromBuffer(kByteBuffer2 + 4));
 }
